requirements.txt
dashscope
fastapi
uvicorn
pyyaml
pathlib
lancedb>=0.6.0
numpy>=1.21.0
pytest>=7.0.0
requests>=2.31.0
websockets==10.4
# 需要 GCC 8.4 以上的版本
